Working FileManager and Viewer prototype and Code refactoring
