+++ .vscode/settings.json
@@ -7,5 +7,5 @@
     { "directory": "./client", "changeProcessCWD": true },
     { "directory": "./server-nest", "changeProcessCWD": true }
   ],
-  "debug.node.autoAttach": "on"
+  "debug.node.autoAttach": "off"
 }
