+++ client/src/components/Search/ProjectCard.vue
