+++ client/src/components/Settings/SettingsEmails.vue
