+++ client/src/pages/pages/search.vue
@@ -83,8 +83,8 @@
 </template>
 
 <script>
-import ProjectCard from '../../components/search/ProjectCard'
-import ProfileCard from '../../components/search/ProfileCard'
+import ProjectCard from '../../components/Search/ProjectCard'
+import ProfileCard from '../../components/Search/ProfileCard'
 import _ from 'lodash'
 import axios from 'axios'
