+++ client/src/pages/settings.vue
@@ -71,10 +71,10 @@
 
 <script>
 import { get } from 'vuex-pathify'
-import SettingsProfile from '../components/settings/settingsProfile.vue'
-import SettingsAccount from '../components/settings/settingsAccount.vue'
-import SettingsEmails from '../components/settings/settingsEmails.vue'
-import SettingsSecurity from '../components/settings/settingsSecurity.vue'
+import SettingsProfile from '../components/Settings/SettingsProfile'
+import SettingsAccount from '../components/Settings/SettingsAccount'
+import SettingsEmails from '../components/Settings/SettingsEmails'
+import SettingsSecurity from '../components/Settings/SettingsSecurity'
 
 export default {
   components: {
